SI-4557 refactor(command): Отдельные файлы для инвокеров
