+++ OOP Patterns/aquarium/test/invokers_test.dart
@@ -2,7 +2,8 @@ import 'package:aquarium/commands/factory/duty_commands_factory.dart';
 import 'package:aquarium/commands/factory/natures_event_factory.dart';
 import 'package:aquarium/fish/fish_factory.dart';
 import 'package:aquarium/fish/subtypes/goldfish.dart';
-import 'package:aquarium/invokers/invoker.dart';
+import 'package:aquarium/invokers/random_invoker.dart';
+import 'package:aquarium/invokers/scheduled_invoker.dart';
 import 'package:aquarium/pool/pool.dart';
 import 'package:aquarium/pool/pool_state.dart';
 import 'package:aquarium/pool/staff.dart';
